Fix Installer, now warns on Errors correctly.

+++ core/update.php
@@ -7,33 +7,20 @@
 echo'<script>/*<![CDATA[*/';
 if(!defined('DS'))define('DS',DIRECTORY_SEPARATOR);
 if(session_status()==PHP_SESSION_NONE)session_start();
-require'db.php';
-$config=$db->query("SELECT * FROM config WHERE id='1'") -> fetch(PDO::FETCH_ASSOC);
-require'sanitise.php';
-function svg($svg,$color=null,$class=null,$size=null){
-	echo'<i class="libre';
-	if($size!=null)echo' libre-'.$size;
-	if($color==true)$svg='col'.DS.$svg;
-	elseif($color!= null)echo' libre-'.$color;
-  if($class!=null)echo' '.$class;
-	echo'">';
-	include'svg'.DS.$svg.'.svg';
-	echo'</i>';
+require_once'db.php';
+$config=$db->query("SELECT * FROM `".$prefix."config` WHERE id='1'")->fetch(PDO::FETCH_ASSOC);
+include'sanitise.php';
+function svg($svg,$class=null,$size=null){
+	echo'<i class="libre'.($size!=null?' libre-'.$size:'').($class!=null?' '.$class:'').'">'.file_get_contents('svg'.DS.$svg.'.svg').'</i>';
 }
-function svg2($svg,$color=null,$class=null,$size=null){
-	$svgout='<i class="libre';
-	if($size!=null)$svgout.=' libre-'.$size;
-	if($color==true)$svg='col'.DS.$svg;
-	elseif($color!=null)$svgout.=' libre-'.$color;
-  if($class!=null)$svgout.=' '.$class;
-	$svgout.='">'.file_get_contents('svg'.DS.$svg.'.svg').'</i>';
-	return$svgout;
+function svg2($svg,$class=null,$size=null){
+	return'<i class="libre'.($size!=null?' libre-'.$size:'').($class!=null?' '.$class:'').'">'.file_get_contents('svg'.DS.$svg.'.svg').'</i>';
 }
 $e='';
 $id=isset($_POST['id'])?filter_input(INPUT_POST,'id',FILTER_SANITIZE_NUMBER_INT):filter_input(INPUT_GET,'id',FILTER_SANITIZE_NUMBER_INT);
 $tbl=isset($_POST['t'])?filter_input(INPUT_POST,'t',FILTER_SANITIZE_STRING):filter_input(INPUT_GET,'t',FILTER_SANITIZE_STRING);
 $col=isset($_POST['c'])?filter_input(INPUT_POST,'c',FILTER_SANITIZE_STRING):filter_input(INPUT_GET,'c',FILTER_SANITIZE_STRING);
-if($tbl=='content'||$tbl=='menu'||$tbl=='config'&&$col=='notes'||$col=='PasswordResetLayout'||$col=='orderEmailLayout'||$col=='orderEmailNotes'||$col=='passwordResetLayout'||$col=='accountActivationLayout'||$col=='bookingEmailLayout'||$col=='bookingAutoReplyLayout'||$col=='contactAutoReplyLayout'||$col=='dateFormat'||$col=='newslettersOptOutLayout'){
+if($tbl=='content'||$tbl=='menu'||$tbl=='config'&&$col=='notes'||$col=='PasswordResetLayout'||$col=='orderEmailLayout'||$col=='orderEmailNotes'||$col=='passwordResetLayout'||$col=='accountActivationLayout'||$col=='bookingEmailLayout'||$col=='bookingAutoReplyLayout'||$col=='contactAutoReplyLayout'||$col=='dateFormat'||$col=='newslettersOptOutLayout'||$col=='php_quicklink'){
   $da=isset($_POST['da'])?filter_input(INPUT_POST,'da',FILTER_UNSAFE_RAW):filter_input(INPUT_GET,'da',FILTER_UNSAFE_RAW);
 }else{
   $da=isset($_POST['da'])?filter_input(INPUT_POST,'da',FILTER_SANITIZE_STRING):filter_input(INPUT_GET,'da',FILTER_SANITIZE_STRING);
@@ -43,71 +30,71 @@ if(strlen($da)<12&&$da=='<p><br></p>')$da=str_replace('<p><br></p>','',$da);
 if(strlen($da)<24&&$da=='%3Cp%3E%3Cbr%3E%3C/p%3E')$da=str_replace('%3Cp%3E%3Cbr%3E%3C/p%3E','',$da);
 $si=session_id();
 $ti=time();
-$s=$db->prepare("SELECT * FROM ".$tbl." WHERE id=:id");
+$s=$db->prepare("SELECT * FROM `".$prefix.$tbl."` WHERE id=:id");
 $s->execute(array(':id'=>$id));
 $r=$s->fetch(PDO::FETCH_ASSOC);
 $oldda=$r[$col];
 if($tbl=='content'&&$col=='status'&&$da=='published'){
-  $q=$db->prepare("UPDATE content SET pti=:pti WHERE id=:id");
+  $q=$db->prepare("UPDATE `".$prefix."content` SET pti=:pti WHERE id=:id");
   $q->execute(
     array(
-      ':pti' => $ti,
-      ':id'  => $id
+      ':pti'=>$ti,
+      ':id' =>$id
     )
   );
 }
 if($tbl=='config'||$tbl=='login'||$tbl=='orders'||$tbl=='orderitems'||$tbl=='messages')$r['contentType']='';
 $log=[
-  'uid'        =>0,
-  'rid'        =>$id,
-  'username'   =>'',
-  'name'       =>'',
-  'view'       =>$r['contentType'],
+  'uid'=>0,
+  'rid'=>$id,
+  'username'=>'',
+  'name'=>'',
+  'view'=>$r['contentType'],
   'contentType'=>$r['contentType'],
-  'refTable'   =>$tbl,
-  'refColumn'  =>$col,
-  'oldda'      =>$oldda,
-  'newda'      =>$da,
-  'action'     =>'update',
-  'ti'         =>$ti
+  'refTable'=>$tbl,
+  'refColumn'=>$col,
+  'oldda'=>$oldda,
+  'newda'=>$da,
+  'action'=>'update',
+  'ti'=>$ti
 ];
 if($r['contentType']=='booking')$log['view']=$r['contentType'].'s';
 if(isset($_SESSION['uid'])){
   $uid=(int)$_SESSION['uid'];
-  $q=$db->prepare("SELECT rank,username,name FROM login WHERE id=:id");
+  $q=$db->prepare("SELECT rank,username,name FROM `".$prefix."login` WHERE id=:id");
   $q->execute(array(':id'=>$uid));
   $u=$q->fetch(PDO::FETCH_ASSOC);
-  if($u['name']!='')$login_user=$u['name'];else$login_user=$u['username'];
-  $log['uid']     =$uid;
+  $login_user=($u['name']!=''?$u['name']:$u['username']);
+  $log['uid']=$uid;
 	$log['username']=$u['username'];
-	$log['name']    =$u['name'];
+	$log['name']=$u['name'];
 }else{
-  $uid       =0;
-  $u['rank'] =0;
+  $uid=0;
+  $u['rank']=0;
   $login_user="Anonymous";
 }
 if($tbl=='login'&&$col=='password'){
-  $da           =password_hash($da,PASSWORD_DEFAULT);
+  $da=password_hash($da,PASSWORD_DEFAULT);
   $log['action']='update password';
-  $log['oldda'] ='';
-  $log['newda'] ='';
+  $log['oldda']='';
+  $log['newda']='';
 }
-if ($tbl=='content'||$tbl=='menu'){
-  $q=$db->prepare("UPDATE $tbl SET eti=:ti,login_user=:login_user,uid=:uid WHERE id=:id");
+if($tbl=='content'||$tbl=='menu'){
+  $q=$db->prepare("UPDATE `".$prefix.$tbl."` SET eti=:ti,login_user=:login_user,uid=:uid WHERE id=:id");
   $q->execute(
     array(
-      'ti'         =>$ti,
-      ':uid'       =>$uid,
+      'ti'=>$ti,
+      ':uid'=>$uid,
       ':login_user'=>$login_user,
-      ':id'        =>$id
+      ':id'=>$id
     )
   );
 }
 if($tbl=='login'&&$col=='username'){
-  $uc1=$db->prepare("SELECT username FROM login WHERE username=:da");
+  $uc1=$db->prepare("SELECT username FROM `".$prefix."login` WHERE username=:da");
   $uc1->execute(array(':da'=>$da));
-  if ($uc1->rowCount()<1){
-    $q=$db->prepare("UPDATE login SET username=:da WHERE id=:id");
+  if($uc1->rowCount()<1){
+    $q=$db->prepare("UPDATE `".$prefix."login` SET username=:da WHERE id=:id");
     $q->execute(
       array(
         ':da'=>$da,
@@ -116,13 +103,13 @@ if($tbl=='login'&&$col=='username'){
     );?>
 	window.top.window.$('#uerror').addClass('hidden');
 <?php }else{
-    $uc2=$db->prepare("SELECT username FROM login WHERE id=:id");
+    $uc2=$db->prepare("SELECT username FROM `".$prefix."login` WHERE id=:id");
     $uc2->execute(array(':id'=>$id));
     $uc=$uc2->fetch(PDO::FETCH_ASSOC);?>
 	window.top.window.$('#uerror').removeClass('hidden');
 <?php }
 }else{
-  $q=$db->prepare("UPDATE $tbl SET $col=:da WHERE id=:id");
+  $q=$db->prepare("UPDATE `".$prefix.$tbl."` SET $col=:da WHERE id=:id");
   $q->execute(
     array(
       ':da'=>$da,
@@ -131,49 +118,48 @@ if($tbl=='login'&&$col=='username'){
   );
 }
 if($tbl=='login'&&$col=='email'){
-	$h=$db->prepare("UPDATE login SET hash=:hash WHERE id=:id");
+	$h=$db->prepare("UPDATE `".$prefix."login` SET hash=:hash WHERE id=:id");
 	$h->execute(
     array(
       ':hash'=>md5($da),
-      ':id'  =>$id
+      ':id'=>$id
     )
   );
 }
 $e=$db->errorInfo();
 if($tbl=='orders'&&$col=='status'&&$da=='archived'){
-  $r=$db->query("SELECT MAX(id) as id FROM orders")->fetch();
+  $r=$db->query("SELECT MAX(id) as id FROM `".$prefix."orders`")->fetch();
   $oid=strtoupper('A').date("ymd",$ti).sprintf("%06d",$r['id']+1,6);
-  $q=$db->prepare("UPDATE orders SET aid=:aid,aid_ti=:aid_ti WHERE id=:id");
+  $q=$db->prepare("UPDATE `".$prefix."orders` SET aid=:aid,aid_ti=:aid_ti WHERE id=:id");
   $q->execute(
     array(
-      ':aid'   =>$oid,
+      ':aid'=>$oid,
       ':aid_ti'=>$ti,
-      ':id'    =>$id
+      ':id'=>$id
     )
   );
 }
 if(is_null($e[2])){
 	if($tbl=='orders'&&$col=='due_ti'){?>
-	window.top.$("#due_ti").val("<?php echo date($config['dateFormat'],$da);?>");
+	window.top.window.$("#due_ti").val("<?php echo date($config['dateFormat'],$da);?>");
 <?php }
 	if($tbl=='content'&&$col=='file'&&$da==''){
-    if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.jpg'))
-      unlink('..'.DS.'media'.DS.'file_'.$id.'.jpg');
-    if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.png'))
-      unlink('..'.DS.'media'.DS.'file_'.$id.'.png');
-    if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.gif'))
-      unlink('..'.DS.'media'.DS.'file_'.$id.'.gif');
-		if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.tif'))
-      unlink('..'.DS.'media'.DS.'file_'.$id.'.tif');
+    if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.jpg'))unlink('..'.DS.'media'.DS.'file_'.$id.'.jpg');
+    if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.png'))unlink('..'.DS.'media'.DS.'file_'.$id.'.png');
+    if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.gif'))unlink('..'.DS.'media'.DS.'file_'.$id.'.gif');
+		if(file_exists('..'.DS.'media'.DS.'file_'.$id.'.tif'))unlink('..'.DS.'media'.DS.'file_'.$id.'.tif');
 	}
+	if($tbl=='config'&&$col=='php_honeypot'){?>
+	window.top.window.$('#php_honeypot_link').html('<?php echo($da!=''?'<a target="_blank" href="'.$da.'">'.$da.'</a>':'Honey Pot File Not Uploaded.');?>');
+<?php }
 	if($tbl=='orderitems'||$tbl=='cart'){
     if($tbl=='cart'&&$col=='quantity'){
       if($da==0){
-        $q=$db->prepare("DELETE FROM cart WHERE id=:id");
+        $q=$db->prepare("DELETE FROM `".$prefix."cart` WHERE id=:id");
         $q->execute(array(':id'=>$id));
         $cnt='';
       }
-      $q=$db->prepare("SELECT SUM(quantity) as quantity FROM cart WHERE si=:si");
+      $q=$db->prepare("SELECT SUM(quantity) as quantity FROM `".$prefix."cart` WHERE si=:si");
       $q->execute(array(':si'=>$si));
       $r=$q->fetch(PDO::FETCH_ASSOC);
       $cnt=$r['quantity'];
@@ -181,28 +167,26 @@ if(is_null($e[2])){
 	window.top.window.$('#cart').html('<?php echo$cnt;?>');
 <?php	}
       if($tbl=='orderitems'){
-        $q=$db->prepare("SELECT oid FROM orderitems WHERE id=:id");
+        $q=$db->prepare("SELECT oid FROM `".$prefix."orderitems` WHERE id=:id");
         $q->execute(array(':id'=>$id));
         $iid=$q->fetch(PDO::FETCH_ASSOC);
       }
       if($tbl=='orderitems'&&$col=='quantity'&&$da==0){
-        $q=$db->prepare("DELETE FROM orderitems WHERE id=:id");
+        $q=$db->prepare("DELETE FROM `".$prefix."orderitems` WHERE id=:id");
         $q->execute(array(':id'=>$id));
       }
-      $total  =0;
-      $content='';
+      $total=0;
+      $content=$html='';
       if($tbl=='cart'){
-        $q=$db->prepare("SELECT * FROM cart WHERE si=:si ORDER BY ti DESC");
+        $q=$db->prepare("SELECT * FROM `".$prefix."cart` WHERE si=:si ORDER BY ti DESC");
         $q->execute(array(':si'=>$si));
       }
       if($tbl=='orderitems'){
-        $q=$db->prepare("SELECT * FROM orderitems WHERE oid=:oid ORDER BY ti ASC,title ASC");
+        $q=$db->prepare("SELECT * FROM `".$prefix."orderitems` WHERE oid=:oid ORDER BY ti ASC,title ASC");
         $q->execute(array(':oid'=>$iid['oid']));
       }
-      $html ='';
-      $total=0;
       while($oi=$q->fetch(PDO::FETCH_ASSOC)){
-        $s=$db->prepare("SELECT * FROM content WHERE id=:id");
+        $s=$db->prepare("SELECT * FROM `".$prefix."content` WHERE id=:id");
         $s->execute(array(':id'=>$oi['iid']));
         $i=$s->fetch(PDO::FETCH_ASSOC);
         $html.='<tr>'.
@@ -228,7 +212,7 @@ if(is_null($e[2])){
                     '<input type="hidden" name="t" value="orderitems">'.
                     '<input type="hidden" name="c" value="quantity">'.
                     '<input type="hidden" name="da" value="0">'.
-                    '<button class="btn btn-default trash">'.svg2('libre-gui-trash',($config['iconsColor']==1?true:null)).'</button>'.
+                    '<button class="btn btn-default trash">'.svg2('libre-gui-trash').'</button>'.
                   '</form>'.
                 '</td>'.
               '</tr>';
@@ -244,11 +228,15 @@ if(is_null($e[2])){
 <?php }
     	if($tbl=='login'&&$col=='gravatar'){
         if($da==''){
-          $sav=$db->prepare("SELECT avatar FROM login WHERE id=:id");
+          $sav=$db->prepare("SELECT avatar FROM `".$prefix."login` WHERE id=:id");
           $sav->execute(array(':id'=>$id));
           $av=$sav->fetch(PDO::FETCH_ASSOC);
-          if($av['avatar']!=''&&file_exists('..'.DS.'media'.DS.'avatar'.DS.$av['avatar']))$avatar='media'.DS.'avatar'.DS.$av['avatar'];else$avatar='images'.DS.'noavatar.jpg';
-        }else$avatar=$da;
+          if($av['avatar']!=''&&file_exists('..'.DS.'media'.DS.'avatar'.DS.$av['avatar']))
+						$avatar='media'.DS.'avatar'.DS.$av['avatar'];
+					else
+						$avatar='images'.DS.'noavatar.jpg';
+        }else
+					$avatar=$da;
 ?>
 	window.top.window.$('#avatar').attr('src','<?php echo$avatar.'?'.time();?>');
 <?php	}
@@ -270,20 +258,20 @@ if(is_null($e[2])){
 	window.top.window.Pace.stop();
 <?php
 echo'/*]]>*/</script>';
-$s=$db->prepare("INSERT INTO logs (uid,rid,username,name,view,contentType,refTable,refColumn,oldda,newda,action,ti) VALUES (:uid,:rid,:username,:name,:view,:contentType,:refTable,:refColumn,:oldda,:newda,:action,:ti)");
+$s=$db->prepare("INSERT INTO `".$prefix."logs` (uid,rid,username,name,view,contentType,refTable,refColumn,oldda,newda,action,ti) VALUES (:uid,:rid,:username,:name,:view,:contentType,:refTable,:refColumn,:oldda,:newda,:action,:ti)");
 $s->execute(
   array(
-    ':uid'        =>$log['uid'],
-    ':rid'        =>$log['rid'],
-    ':username'   =>$log['username'],
-    ':name'       =>$log['name'],
-    ':view'       =>$log['view'],
+    ':uid'=>$log['uid'],
+    ':rid'=>$log['rid'],
+    ':username'=>$log['username'],
+    ':name'=>$log['name'],
+    ':view'=>$log['view'],
     ':contentType'=>$log['contentType'],
-    ':refTable'   =>$log['refTable'],
-    ':refColumn'  =>$log['refColumn'],
-    ':oldda'      =>$log['oldda'],
-    ':newda'      =>$log['newda'],
-    ':action'     =>$log['action'],
-    ':ti'         =>$log['ti']
+    ':refTable'=>$log['refTable'],
+    ':refColumn'=>$log['refColumn'],
+    ':oldda'=>$log['oldda'],
+    ':newda'=>$log['newda'],
+    ':action'=>$log['action'],
+    ':ti'=>$log['ti']
   )
 );
